requirements.txt
faiss-cpu
groq
langchain-groq
pyPDF2
langchain_google_genai
langchain
streamlit
langchain-community
python-dotenv
pypdf
duckduckgo-search